requirements.txt
griptape
python-dotenv

